pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mycosoft-mas"
version = "1.0.0"
description = "Mycosoft Multi-Agent System"
authors = ["Mycosoft <info@mycosoft.com>"]
readme = "README.md"
packages = [{include = "mycosoft_mas"}]

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.1"
uvicorn = "^0.23.2"
pydantic = "^2.1.1"
aiohttp = "^3.9.1"
aioredis = "^2.0.1"
prometheus-client = "^0.17.1"
pyyaml = "^6.0.1"
selenium = "^4.15.2"
redis = "^5.0.1"
sqlalchemy = "^2.0.23"
alembic = "^1.11.1"
celery = "^5.2.7"
flower = "^2.0.1"
aiosqlite = "^0.19.0"
aiosmtplib = "^2.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
httpx = "~0.24.1"
jinja2 = "^3.1.2"
python-dotenv = "^1.0.0"
asyncio = "^3.4.3"
python-bitcoinlib = "^0.12.2"
transformers = {version = "^4.35.2", extras = ["torch"], markers = "platform_machine != 'aarch64'"}
torch = {version = "^2.2.0+cpu", source = "pytorch"}
sentence-transformers = "^2.2.2"
scikit-learn = "^1.3.2"
opentelemetry-api = "^1.18.0"
opentelemetry-sdk = "^1.18.0"
opentelemetry-instrumentation-fastapi = "^0.39b0"
opentelemetry-exporter-otlp = "^1.18.0"
requests = "^2.31.0"
python-telegram-bot = "~20.4"
slack-sdk = "^3.26.1"
python-docx = "^1.1.2"
pypdf2 = "^3.0.1"
stripe = "^12.0.0"
opencv-python-headless = "^4.11.0.86"
undetected-chromedriver = "^3.5.5"
pywinauto = "^0.6.9"
docker = "^7.0.0"
aiofiles = "^23.2.1"
pandas = "^2.2.1"
numpy = "^1.26.4"
matplotlib = "^3.8.3"
rdflib = "^7.0.0"
semver = "^3.0.2"
spacy = "^3.7.4"
web3 = "^6.15.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.4.1"
flake8 = "^6.1.0"
pre-commit = "^3.3.3"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true 